Changes All Repo

+++ src/main/java/com/bbots/gl/repository/GlAttributesRepository.java
@@ -19,7 +19,7 @@ public class GlAttributesRepository {
 	
 	
 	private final RowMapper<GlAttributes> glAttributes = (rs, rowNum) -> new GlAttributes(
-			 rs.getInt("ORGCODE"),
+			 rs.getLong("ORGCODE"),
 			 rs.getInt("GLNO"),
 			 rs.getString("GLATTRID"),
 			 rs.getString("GLATTRVALUE"),
